Removed subject from NATSClient, implemented instead in respective sender/receiver
